move create_manifest to cli file
